Added icon selected on mouse over

+++ Library/OcBootManagementLib/DefaultEntryChoice.c
@@ -900,7 +900,7 @@ OcSetDefaultBootEntry (
     }
 
     LoadOption->Attributes         = LOAD_OPTION_ACTIVE | LOAD_OPTION_CATEGORY_BOOT;
-    LoadOption->FilePathListLength = DevicePathSize;
+    LoadOption->FilePathListLength = (UINT16) DevicePathSize;
     CopyMem (LoadOption + 1, Entry->Name, LoadOptionNameSize);
     CopyMem ((UINT8 *) (LoadOption + 1) + LoadOptionNameSize, Entry->DevicePath, DevicePathSize);
